refactor(cross-entity): extract sub-components into CrossEntityAutomation folder
Extracted 11 components from CrossEntityAutomationView.tsx into dedicated files in src/components/CrossEntityAutomation/:

- OperationBadge.tsx — operation type badge with semantic colors
- TypeBadge.tsx — automation type badge with icons
- DetectionCoverageBanner.tsx — coverage notice banner
- FieldPills.tsx — field match visualization pills
- FieldMatchVerdict.tsx — field match verdict block with hit/miss styling
- ChildEntitySection.tsx — nested downstream entity pipeline section
- StepBlock.tsx — individual pipeline step with branch visualization
- TracePipeline.tsx — numbered step list for one trace
- MessagePipelineSteps.tsx — steps for message pipeline without entry context
- EntityMessagePipelineRow.tsx — entity with automation but no inbound entry points
- EntityPipelineRow.tsx — entity with inbound entry points (most complex)

Each component is self-contained with its own makeStyles hook. Shared utilities (entityColor, ENTITY_COLORS) kept in main file. Created index.ts barrel export.

Main CrossEntityAutomationView.tsx now imports from ./CrossEntityAutomation and delegates rendering to extracted components.

File reduction: 1616 lines → 849 lines in main component file.
