+++ src/components/CrossEntityAutomationView.tsx
@@ -1,18 +1,15 @@
-import { useState, useCallback, Fragment, type ReactElement } from 'react';
+import { useState, useCallback, Fragment } from 'react';
 import {
   Text,
   Title3,
   Card,
   Badge,
   makeStyles,
-  mergeClasses,
   tokens,
   Tab,
   TabList,
   SelectTabData,
   SelectTabEvent,
-  Tooltip,
-  Button,
   ToggleButton,
   Checkbox,
 } from '@fluentui/react-components';
@@ -20,28 +17,17 @@ import { FilterBar, FilterGroup } from './FilterBar';
 import {
   ArrowRight24Regular,
   Warning24Regular,
-  Warning20Regular,
-  Lightbulb24Regular,
-  CheckmarkCircle16Regular,
-  ErrorCircle16Regular,
   Info16Regular,
 } from '@fluentui/react-icons';
 import { EmptyState } from './EmptyState';
 import {
-  PluginsIcon as BracesVariable24Regular,
-  FlowsIcon as CloudFlow24Regular,
-  BusinessRulesIcon as ClipboardTaskListLtr24Regular,
-  ClassicWorkflowsIcon as ClipboardSettings24Regular,
-  WebResourcesIcon as DocumentGlobe24Regular,
-} from './componentIcons';
+  DetectionCoverageBanner,
+  EntityPipelineRow,
+  EntityMessagePipelineRow,
+  OperationBadge,
+} from './CrossEntityAutomation';
 import type {
   CrossEntityAnalysisResult,
-  CrossEntityTrace,
-  AutomationActivation,
-  CrossEntityEntityView,
-  EntityAutomationPipeline,
-  MessagePipeline,
-  PipelineStep,
 } from '../core';
 import type { EntityBlueprint } from '../core';
 
@@ -58,22 +44,7 @@ const ENTITY_COLORS = [
   '#0078d4', '#107c10', '#ca5010', '#8764b8',
   '#038387', '#c239b3', '#e3008c', '#004b50',
 ];
-function entityColor(index: number) { return ENTITY_COLORS[index % ENTITY_COLORS.length]; }
-
-/* ─────────────────────────────────────────────────────────────────────────
-   Type helpers
-───────────────────────────────────────────────────────────────────────── */
-// tokens.fontSizeBase200 = 12px — used for inline type icon sizing (AUDIT-004 compliant)
-const TYPE_ICON_STYLE = { width: tokens.fontSizeBase200, height: tokens.fontSizeBase200, flexShrink: 0, verticalAlign: 'middle' } as const;
-function typeIcon(type: AutomationActivation['automationType'] | PipelineStep['automationType']): ReactElement {
-  switch (type) {
-    case 'Plugin': return <BracesVariable24Regular style={TYPE_ICON_STYLE} />;
-    case 'Flow': return <CloudFlow24Regular style={TYPE_ICON_STYLE} />;
-    case 'BusinessRule': return <ClipboardTaskListLtr24Regular style={TYPE_ICON_STYLE} />;
-    case 'ClassicWorkflow': return <ClipboardSettings24Regular style={TYPE_ICON_STYLE} />;
-    default: return <ArrowRight24Regular style={TYPE_ICON_STYLE} />;
-  }
-}
+function entityColor(index: number): string { return ENTITY_COLORS[index % ENTITY_COLORS.length]; }
 
 /* ─────────────────────────────────────────────────────────────────────────
    Styles
@@ -874,772 +845,5 @@ export function CrossEntityAutomationView({ analysis }: CrossEntityAutomationVie
         </div>
       )}
     </div>
-  );
-}
-
-/* ─────────────────────────────────────────────────────────────────────────
-   EntityPipelineRow — one collapsible row per target entity (with entry points)
-───────────────────────────────────────────────────────────────────────── */
-function EntityPipelineRow({
-  logicalName, view, pipeline, color, expanded, onToggle, analysis,
-}: {
-  logicalName: string;
-  view: CrossEntityEntityView;
-  pipeline: EntityAutomationPipeline;
-  color: string;
-  expanded: boolean;
-  onToggle: () => void;
-  analysis: CrossEntityAnalysisResult;
-}) {
-  const styles = useStyles();
-  const totalSteps = pipeline.messagePipelines.reduce((sum, mp) => sum + mp.steps.length, 0);
-
-  return (
-    <div className={styles.entityRow}>
-      {/* Header */}
-      <div
-        className={mergeClasses(styles.entityHeader, expanded && styles.entityHeaderOpen)}
-        role="button"
-        tabIndex={0}
-        aria-expanded={expanded}
-        onClick={onToggle}
-        onKeyDown={(e) => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); onToggle(); } }}
-        style={{ borderLeft: `4px solid ${color}` }}
-      >
-        <div className={styles.entityInfo}>
-          <div className={styles.entityNameRow}>
-            <Text weight="semibold">{view.entityDisplayName}</Text>
-            <Text className={styles.entityLogicalName}>
-              {logicalName}
-            </Text>
-            {pipeline.messagePipelines.map(mp => (
-              <OperationBadge key={mp.message} operation={mp.message} />
-            ))}
-            {pipeline.hasCrossEntityOutput && (
-              <Badge appearance="tint" shape="rounded" color="informative">&rarr; cross-entity</Badge>
-            )}
-            {pipeline.hasExternalInteraction && (
-              <Badge appearance="tint" shape="rounded" color="brand">&#8627; External calls</Badge>
-            )}
-            {view.traces.length > 1 && (
-              <Badge appearance="tint" shape="rounded" color="informative">{view.traces.length} entry points</Badge>
-            )}
-          </div>
-
-
-        </div>
-
-        <Text className={styles.entityHeaderCount}>
-          {expanded ? '▲' : '▼'} {totalSteps}
-        </Text>
-      </div>
-
-      {/* Expanded body — pipeline-first: one section per message pipeline, inbound triggers as context */}
-      {expanded && (
-        <div className={styles.entityBody} style={{ borderLeft: `4px solid ${color}40` }}>
-          {pipeline.messagePipelines.map((mp, mpi) => {
-            // Inbound entry points from other entities that trigger this message on this entity
-            const inboundTraces = view.traces.filter(t => t.entryPoint.operation === mp.message);
-            // IDs of steps activated by at least one inbound entry point (WillFire or WillFireNoFilter)
-            const triggeredStepIds = new Set(
-              inboundTraces.flatMap(t =>
-                t.activations
-                  .filter(a => a.firingStatus !== 'WontFire')
-                  .map(a => a.automationId)
-              )
-            );
-            return (
-              <div key={mp.message}>
-                {mpi > 0 && <div className={styles.traceDivider} />}
-                <div className={styles.traceSubHeader}>
-                  <OperationBadge operation={mp.message} /> pipeline
-                </div>
-                <MessagePipelineSteps
-                  mp={mp}
-                  analysis={analysis}
-                  color={color}
-                  entityDisplayName={view.entityDisplayName}
-                  triggeredStepIds={triggeredStepIds.size > 0 ? triggeredStepIds : undefined}
-                  inboundTraces={inboundTraces.length > 0 ? inboundTraces : undefined}
-                />
-              </div>
-            );
-          })}
-
-          {/* Custom Action Triggers — entry points with operation === 'Action' are never matched
-              by any MessagePipeline.message ('Create'|'Update'|'Delete'|'Manual'), so they must
-              be surfaced in a dedicated section. */}
-          {(() => {
-            const actionTraces = view.traces.filter(t => t.entryPoint.operation === 'Action');
-            if (actionTraces.length === 0) return null;
-            return (
-              <>
-                <div className={styles.traceDivider} />
-                <div className={styles.traceSubHeader}>
-                  <OperationBadge operation="Action" /> Custom Action triggers
-                </div>
-                {actionTraces.map((t, i) => (
-                  <div key={`action-${i}`} className={styles.actionInfoCard}>
-                    <div className={styles.actionInfoCardHeader}>
-                      {typeIcon(t.entryPoint.automationType)}
-                      <Text className={styles.actionInfoCardName}>{t.entryPoint.automationName}</Text>
-                      <Badge appearance="outline" shape="rounded" color="informative" className={styles.tinyBadge}>
-                        {t.entryPoint.automationType}
-                      </Badge>
-                    </div>
-                    <Text className={styles.actionInfoCardSource}>
-                      Source entity: {t.entryPoint.sourceEntityDisplayName}
-                    </Text>
-                    {t.entryPoint.customActionApiName && (
-                      <Text className={styles.actionInfoCardApiName}>
-                        via <code>{t.entryPoint.customActionApiName}</code>
-                      </Text>
-                    )}
-                    <div className={styles.actionInfoCardNote}>
-                      <Info16Regular className={styles.actionInfoCardNoteIcon} />
-                      <Text className={styles.actionInfoCardNoteText}>
-                        Internal effects on this entity&apos;s pipeline are unknown — custom action
-                        logic is not statically analysed.
-                      </Text>
-                    </div>
-                  </div>
-                ))}
-              </>
-            );
-          })()}
-        </div>
-      )}
-    </div>
-  );
-}
-
-/* ─────────────────────────────────────────────────────────────────────────
-   EntityMessagePipelineRow — entity with automation but NO inbound entry points
-───────────────────────────────────────────────────────────────────────── */
-function EntityMessagePipelineRow({
-  pipeline, color, expanded, onToggle, analysis,
-}: {
-  pipeline: EntityAutomationPipeline;
-  color: string;
-  expanded: boolean;
-  onToggle: () => void;
-  analysis: CrossEntityAnalysisResult;
-}) {
-  const styles = useStyles();
-  const totalSteps = pipeline.messagePipelines.reduce((sum, mp) => sum + mp.steps.length, 0);
-
-  return (
-    <div className={styles.entityRow}>
-      <div
-        className={mergeClasses(styles.entityHeader, expanded && styles.entityHeaderOpen)}
-        role="button"
-        tabIndex={0}
-        aria-expanded={expanded}
-        onClick={onToggle}
-        onKeyDown={(e) => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); onToggle(); } }}
-        style={{ borderLeft: `4px solid ${color}` }}
-      >
-        <div className={styles.entityInfo}>
-          <div className={styles.entityNameRow}>
-            <Text weight="semibold">{pipeline.entityDisplayName}</Text>
-            <Text className={styles.entityLogicalName}>
-              {pipeline.entityLogicalName}
-            </Text>
-            {pipeline.messagePipelines.map(mp => (
-              <OperationBadge key={mp.message} operation={mp.message} />
-            ))}
-            {pipeline.hasCrossEntityOutput && (
-              <Badge appearance="tint" shape="rounded" color="informative">&rarr; cross-entity</Badge>
-            )}
-            {pipeline.hasExternalInteraction && (
-              <Badge appearance="tint" shape="rounded" color="brand">&#8627; External calls</Badge>
-            )}
-          </div>
-
-        </div>
-
-        <Text className={styles.entityHeaderCount}>
-          {expanded ? '▲' : '▼'} {totalSteps}
-        </Text>
-      </div>
-
-      {expanded && (
-        <div className={styles.entityBody} style={{ borderLeft: `4px solid ${color}40` }}>
-          {pipeline.messagePipelines.map((mp, mpi) => (
-            <div key={mp.message}>
-              {pipeline.messagePipelines.length > 1 && (
-                <div className={styles.traceSubHeader}>
-                  <OperationBadge operation={mp.message} /> pipeline
-                </div>
-              )}
-              <MessagePipelineSteps mp={mp} analysis={analysis} color={color} entityDisplayName={pipeline.entityDisplayName} />
-              {mpi < pipeline.messagePipelines.length - 1 && (
-                <div className={styles.traceDivider} />
-              )}
-            </div>
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
-
-/* ─────────────────────────────────────────────────────────────────────────
-   MessagePipelineSteps — steps for a message pipeline without entry-point context
-───────────────────────────────────────────────────────────────────────── */
-function MessagePipelineSteps({
-  mp, analysis, color, entityDisplayName, triggeredStepIds, inboundTraces,
-}: {
-  mp: MessagePipeline;
-  analysis: CrossEntityAnalysisResult;
-  color: string;
-  entityDisplayName: string;
-  /** IDs of steps activated by an inbound cross-entity entry point — shown with ← inbound badge */
-  triggeredStepIds?: Set<string>;
-  /** Inbound traces for this pipeline — used in tooltip on the ← inbound badge */
-  inboundTraces?: CrossEntityTrace[];
-}) {
-  const styles = useStyles();
-
-  return (
-    <div>
-      {mp.steps.map((step, i) => {
-        const hasDownstream = !!step.downstream;
-        const downstreamView = hasDownstream
-          ? analysis.entityViews.get(step.downstream!.targetEntity) ?? null
-          : null;
-        const downstreamColor = hasDownstream
-          ? entityColor(Array.from(analysis.allEntityPipelines.keys()).indexOf(step.downstream!.targetEntity))
-          : color;
-
-        return (
-          <div key={step.automationId + i}>
-            <div className={styles.pipelineStepRow}>
-              <div className={mergeClasses(styles.stepMain, hasDownstream ? styles.stepMainHasBranch : undefined)}>
-                <span className={styles.stepNum}>{i + 1}</span>
-                {triggeredStepIds?.has(step.automationId) && (
-                  <Tooltip
-                    content={(inboundTraces ?? []).map(t => `← ${t.entryPoint.automationName} (from ${t.entryPoint.sourceEntityDisplayName})`).join(' · ')}
-                    relationship="description"
-                  >
-                    <Badge appearance="filled" shape="rounded" color="informative" className={styles.stepBadge}>
-                      &larr; inbound
-                    </Badge>
-                  </Tooltip>
-                )}
-                <TypeBadge type={step.automationType} />
-                <span className={styles.stepName}>{step.automationName}</span>
-                {step.stageName && <span className={styles.stepStage}>{step.stageName}</span>}
-                {step.rank !== undefined && <span className={styles.stepStage}>#{step.rank}</span>}
-                <Badge appearance="tint" shape="rounded" color={step.mode === 'Sync' ? 'warning' : 'success'} className={styles.stepBadge}>
-                  {step.mode}
-                </Badge>
-                {step.firesForAllUpdates && (
-                  <Tooltip content="No filtering attributes — fires on ALL updates" relationship="description">
-                    <Badge appearance="filled" shape="rounded" color="danger" className={styles.stepBadge}>&#9888; No filter</Badge>
-                  </Tooltip>
-                )}
-                {step.hasExternalCalls && (
-                  <Badge appearance="tint" shape="rounded" color="brand" className={styles.stepBadge}>
-                    External calls
-                  </Badge>
-                )}
-                {step.filteringAttributes.length > 0 && !step.firesForAllUpdates && mp.message === 'Update' && (
-                  <Text className={styles.stepFilterText}>
-                    filters: {step.filteringAttributes.slice(0, 3).join(', ')}
-                    {step.filteringAttributes.length > 3 && ` +${step.filteringAttributes.length - 3}`}
-                  </Text>
-                )}
-              </div>
-
-              {hasDownstream && step.downstream && (
-                <div className={styles.branchBlock} style={{ borderLeftColor: downstreamColor }}>
-                  <div className={styles.branchTarget} style={{ color: downstreamColor }}>
-                    &rarr; {step.downstream.targetEntityDisplayName}
-                    <OperationBadge operation={step.downstream.operation} />
-                  </div>
-                  {step.downstream.fields.length > 0 && (
-                    <div className={styles.branchFields}>
-                      {step.downstream.fields.slice(0, 4).map(f => (
-                        <span key={f} className={styles.branchField}>{f}</span>
-                      ))}
-                      {step.downstream.fields.length > 4 && (
-                        <span className={styles.branchField}>+{step.downstream.fields.length - 4}</span>
-                      )}
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
-
-            {/* Connectors and external calls for flow steps */}
-            {step.connectionReferences && step.connectionReferences.length > 0 && (
-              <div className={styles.branchFields} style={{ marginTop: tokens.spacingVerticalXS }}>
-                {step.connectionReferences.map((ref, ri) => (
-                  <Badge key={ri} appearance="tint" shape="rounded" size="small" color="subtle">
-                    {ref}
-                  </Badge>
-                ))}
-              </div>
-            )}
-            {step.externalCallSummaries && step.externalCallSummaries.length > 0 && (
-              <div style={{ marginTop: tokens.spacingVerticalXS, display: 'flex', flexDirection: 'column', gap: tokens.spacingVerticalXXS }}>
-                {step.externalCallSummaries.map((call, ci) => (
-                  <Text key={ci} className={styles.stepFilterText} style={{ wordBreak: 'break-all' }}>
-                    {call.method ? `${call.method} ` : ''}{call.url || call.domain}
-                  </Text>
-                ))}
-              </div>
-            )}
-
-            {/* Inline child pipeline (depth 0 only for message pipeline rows) */}
-            {downstreamView && (
-              <ChildEntitySection
-                entityView={downstreamView}
-                trace={downstreamView.traces[0]}
-                analysis={analysis}
-                depth={1}
-                accentColor={downstreamColor}
-                parentDisplayName={entityDisplayName}
-              />
-            )}
-          </div>
-        );
-      })}
-
-      {mp.steps.length === 0 && (
-        <Text className={styles.emptyPipelineText}>
-          No automations registered for {mp.message}.
-        </Text>
-      )}
-    </div>
-  );
-}
-
-/* ─────────────────────────────────────────────────────────────────────────
-   TracePipeline — the numbered step list for one trace
-───────────────────────────────────────────────────────────────────────── */
-function TracePipeline({
-  trace, analysis, depth, parentEntityDisplayName,
-}: {
-  trace: CrossEntityTrace;
-  analysis: CrossEntityAnalysisResult;
-  depth: number;
-  parentEntityDisplayName: string;
-}) {
-  const styles = useStyles();
-  const [showWontFire, setShowWontFire] = useState(false);
-
-  const willFire = trace.activations.filter(a => a.firingStatus !== 'WontFire');
-  const wontFire = trace.activations.filter(a => a.firingStatus === 'WontFire');
-
-  return (
-    <div>
-      {/* Custom action informational note */}
-      {trace.entryPoint.operation === 'Action' && (
-        <div className={styles.entryFieldsRow} style={{ display: 'flex', alignItems: 'flex-start', gap: tokens.spacingHorizontalXS }}>
-          <Info16Regular style={{ color: tokens.colorBrandForeground1, flexShrink: 0, marginTop: '2px' }} />
-          <Text size={200} style={{ fontStyle: 'italic' }}>
-            This entry point calls a bound custom action
-            {trace.entryPoint.customActionApiName ? ` (${trace.entryPoint.customActionApiName})` : ''}.
-            The action&apos;s internal effects on this entity are not analysed — plugins registered on the
-            action&apos;s message may also fire.
-          </Text>
-        </div>
-      )}
-
-      {/* Numbered steps that will fire */}
-      {willFire.map((act, i) => (
-        <StepBlock
-          key={act.automationId + i}
-          stepNum={i + 1}
-          activation={act}
-          trace={trace}
-          analysis={analysis}
-          depth={depth}
-          parentEntityDisplayName={parentEntityDisplayName}
-        />
-      ))}
-
-      {willFire.length === 0 && (
-        <Text className={styles.emptyPipelineText}>
-          {trace.entryPoint.operation === 'Action'
-            ? 'No CRUD-registered automations detected for this custom action call. See note above.'
-            : 'No automations registered on this entity for this message.'}
-        </Text>
-      )}
-
-      {/* Won't fire — collapsed section */}
-      {wontFire.length > 0 && (
-        <>
-          <div className={styles.wontFireBtnWrapper}>
-            <Button appearance="subtle" size="small" onClick={() => setShowWontFire(s => !s)}>
-              {showWontFire ? '▲' : '▼'} {wontFire.length} automation{wontFire.length !== 1 ? 's' : ''} won't fire for this entry point
-            </Button>
-          </div>
-          {showWontFire && wontFire.map((act, i) => (
-            <div key={i} className={styles.wontFireItem}>
-              <span>{typeIcon(act.automationType)}</span>
-              <Badge appearance="outline" shape="rounded" color="informative" className={styles.tinyBadge}>{act.automationType}</Badge>
-              <Text className={styles.wontFireItemName}>
-                {act.automationName}
-              </Text>
-              {act.stageName && (
-                <Text className={styles.wontFireItemStage}>{act.stageName}</Text>
-              )}
-              {act.filteringAttributes.length > 0 && (
-                <Text className={styles.wontFireItemFilter}>
-                  — filter: {act.filteringAttributes.join(', ')} (not in [{trace.entryPoint.fields.join(', ')}])
-                </Text>
-              )}
-            </div>
-          ))}
-        </>
-      )}
-
-      {/* Entry point field context */}
-      {trace.entryPoint.fields.length > 0 && (
-        <div className={styles.entryFieldsRow}>
-          Entry fields: {trace.entryPoint.fields.slice(0, 8).map(f => (
-            <span key={f} className={styles.entryFieldPill}>{f}</span>
-          ))}
-          {trace.entryPoint.fields.length > 8 && <span>+{trace.entryPoint.fields.length - 8} more</span>}
-        </div>
-      )}
-    </div>
-  );
-}
-
-/* ─────────────────────────────────────────────────────────────────────────
-   StepBlock — one numbered activation row + field match verdict + optional child pipeline
-───────────────────────────────────────────────────────────────────────── */
-function StepBlock({
-  stepNum, activation, trace, analysis, depth, parentEntityDisplayName,
-}: {
-  stepNum: number;
-  activation: AutomationActivation;
-  trace: CrossEntityTrace;
-  analysis: CrossEntityAnalysisResult;
-  depth: number;
-  parentEntityDisplayName: string;
-}) {
-  const styles = useStyles();
-  const hasDownstream = !!activation.downstream;
-
-  // Find downstream entity view and its relevant trace (max depth 2)
-  const downstreamView =
-    hasDownstream && depth < 2
-      ? analysis.entityViews.get(activation.downstream!.targetEntity)
-      : null;
-
-  const downstreamTrace = downstreamView
-    ? (downstreamView.traces.find(t => t.entryPoint.automationId === activation.automationId) ?? downstreamView.traces[0])
-    : null;
-
-  // Downstream entity accent color (cycle from a shifted palette index)
-  const downstreamColor = entityColor(
-    Array.from(analysis.entityViews.keys()).indexOf(activation.downstream?.targetEntity ?? '') % ENTITY_COLORS.length
-  );
-
-  return (
-    <div>
-      {/* Step row: main block + optional branch */}
-      <div className={styles.pipelineStepRow}>
-        <div className={mergeClasses(styles.stepMain, hasDownstream ? styles.stepMainHasBranch : undefined)}>
-          <span className={styles.stepNum}>{stepNum}</span>
-          <TypeBadge type={activation.automationType} />
-          <span className={styles.stepName}>{activation.automationName}</span>
-          {activation.stageName && (
-            <span className={styles.stepStage}>{activation.stageName}</span>
-          )}
-          {activation.rank !== undefined && (
-            <span className={styles.stepStage}>#{activation.rank}</span>
-          )}
-          <Badge appearance="tint" shape="rounded" color={activation.mode === 'Sync' ? 'warning' : 'success'} className={styles.stepBadge}>
-            {activation.mode}
-          </Badge>
-          {activation.firingStatus === 'WillFireNoFilter' && (
-            <Tooltip content="No filtering attributes — fires on ALL updates" relationship="description">
-              <Badge appearance="filled" shape="rounded" color="danger" className={styles.stepBadge}>&#9888; No filter</Badge>
-            </Tooltip>
-          )}
-        </div>
-
-        {/* Branch block */}
-        {hasDownstream && activation.downstream && (
-          <div className={styles.branchBlock} style={{ borderLeftColor: downstreamColor }}>
-            <div className={styles.branchTarget} style={{ color: downstreamColor }}>
-              &rarr; {activation.downstream.targetEntityDisplayName}
-              <OperationBadge operation={activation.downstream.operation} />
-            </div>
-            {activation.downstream.fields.length > 0 && (
-              <div className={styles.branchFields}>
-                {activation.downstream.fields.slice(0, 4).map(f => (
-                  <span key={f} className={styles.branchField}>{f}</span>
-                ))}
-                {activation.downstream.fields.length > 4 && (
-                  <span className={styles.branchField}>+{activation.downstream.fields.length - 4}</span>
-                )}
-              </div>
-            )}
-          </div>
-        )}
-      </div>
-
-      {/* Field match verdict */}
-      <FieldMatchVerdict activation={activation} trace={trace} styles={styles} />
-
-      {/* Inline downstream pipeline */}
-      {downstreamView && downstreamTrace && depth < 2 && (
-        <ChildEntitySection
-          entityView={downstreamView}
-          trace={downstreamTrace}
-          analysis={analysis}
-          depth={depth + 1}
-          accentColor={downstreamColor}
-          parentDisplayName={parentEntityDisplayName}
-        />
-      )}
-    </div>
-  );
-}
-
-/* ─────────────────────────────────────────────────────────────────────────
-   ChildEntitySection — indented inline downstream pipeline
-───────────────────────────────────────────────────────────────────────── */
-function ChildEntitySection({
-  entityView, trace, analysis, depth, accentColor, parentDisplayName,
-}: {
-  entityView: CrossEntityEntityView;
-  trace: CrossEntityTrace;
-  analysis: CrossEntityAnalysisResult;
-  depth: number;
-  accentColor: string;
-  parentDisplayName: string;
-}) {
-  const styles = useStyles();
-  const [open, setOpen] = useState(true);
-
-  const willFire = trace.activations.filter(a => a.firingStatus !== 'WontFire');
-  const stepCount = trace.activations.length;
-
-  return (
-    <div className={styles.childSection}>
-      <div
-        className={styles.childHeader}
-        role="button"
-        tabIndex={0}
-        aria-expanded={open}
-        style={{ borderLeft: `3px solid ${accentColor}`, borderColor: `${accentColor}` }}
-        onClick={() => setOpen(o => !o)}
-        onKeyDown={(e) => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); setOpen(o => !o); } }}
-      >
-        <span className={styles.childArrow} style={{ color: accentColor }}>&#8627;</span>
-        <Text weight="semibold" className={styles.childEntityName}>{entityView.entityDisplayName}</Text>
-        <Text className={styles.childLogicalName}>
-          {entityView.entityLogicalName}
-        </Text>
-        <OperationBadge operation={trace.entryPoint.operation} />
-        <Text className={styles.childStepCount}>
-          {willFire.length} step{willFire.length !== 1 ? 's' : ''} &bull; {open ? '▲' : `▶ expand (${stepCount})`}
-        </Text>
-      </div>
-
-      {open && (
-        <div className={styles.childSteps} style={{ borderLeft: `3px solid ${accentColor}40` }}>
-          <TracePipeline
-            trace={trace}
-            analysis={analysis}
-            depth={depth}
-            parentEntityDisplayName={entityView.entityDisplayName}
-          />
-        </div>
-      )}
-
-      <div className={styles.returnMarker} style={{ borderLeft: `2px solid ${accentColor}40` }}>
-        &#8617; back to {parentDisplayName}
-      </div>
-    </div>
-  );
-}
-
-/* ─────────────────────────────────────────────────────────────────────────
-   FieldMatchVerdict — verdict block shown beneath each step
-───────────────────────────────────────────────────────────────────────── */
-function FieldMatchVerdict({
-  activation, trace, styles,
-}: {
-  activation: AutomationActivation;
-  trace: CrossEntityTrace;
-  styles: ReturnType<typeof useStyles>;
-}) {
-  const { firingStatus, matchedFields, filteringAttributes } = activation;
-  const isCreateOrDelete = trace.entryPoint.operation === 'Create' || trace.entryPoint.operation === 'Delete';
-
-  if (firingStatus === 'WillFire') {
-    if (isCreateOrDelete) {
-      return (
-        <div className={mergeClasses(styles.fieldMatchBase, styles.fieldMatchFires)}>
-          <div className={styles.matchVerdict} style={{ color: tokens.colorPaletteGreenForeground1 }}>
-            <CheckmarkCircle16Regular /> FIRES — all {trace.entryPoint.operation} automations fire
-          </div>
-        </div>
-      );
-    }
-    return (
-      <div className={mergeClasses(styles.fieldMatchBase, styles.fieldMatchFires)}>
-        <div className={styles.matchVerdict} style={{ color: tokens.colorPaletteGreenForeground1 }}>
-          <CheckmarkCircle16Regular /> WILL FIRE
-          {filteringAttributes.length > 0 && ' — matched filtering attributes'}
-        </div>
-        {matchedFields.length > 0 && (
-          <FieldPills
-            entryFields={trace.entryPoint.fields}
-            filterFields={filteringAttributes}
-            matchedFields={matchedFields}
-            styles={styles}
-          />
-        )}
-      </div>
-    );
-  }
-
-  if (firingStatus === 'WontFire') {
-    return (
-      <div className={mergeClasses(styles.fieldMatchBase, styles.fieldMatchNoFire)}>
-        <div className={styles.matchVerdict} style={{ color: tokens.colorNeutralForeground3 }}>
-          <ErrorCircle16Regular /> WON'T FIRE — no overlap between updated fields and filter
-        </div>
-        <FieldPills
-          entryFields={trace.entryPoint.fields}
-          filterFields={filteringAttributes}
-          matchedFields={matchedFields}
-          styles={styles}
-        />
-      </div>
-    );
-  }
-
-  if (firingStatus === 'WillFireNoFilter') {
-    return (
-      <div className={mergeClasses(styles.fieldMatchBase, styles.fieldMatchNoFilter)}>
-        <div className={styles.matchVerdict} style={{ color: tokens.colorPaletteRedForeground1, display: 'flex', alignItems: 'center', gap: tokens.spacingHorizontalXS }}>
-          <Warning20Regular /> WILL FIRE — no filtering attributes, fires on ALL updates
-        </div>
-        <Text className={styles.noFilterAdvisory}>
-          Add filtering attributes to this {activation.automationType.toLowerCase()} to improve performance.
-        </Text>
-      </div>
-    );
-  }
-
-  return null;
-}
-
-/* ─────────────────────────────────────────────────────────────────────────
-   FieldPills — entry fields vs filter fields with hit/miss highlighting
-───────────────────────────────────────────────────────────────────────── */
-function FieldPills({
-  entryFields, filterFields, matchedFields, styles,
-}: {
-  entryFields: string[];
-  filterFields: string[];
-  matchedFields: string[];
-  styles: ReturnType<typeof useStyles>;
-}) {
-  const matchedSet = new Set(matchedFields.map(f => f.toLowerCase()));
-
-  const showEntry = entryFields.length > 0;
-  const showFilter = filterFields.length > 0;
-
-  return (
-    <div>
-      {showEntry && (
-        <div className={styles.fieldPillRow}>
-          <Text className={styles.fieldPillLabel}>Updated:</Text>
-          {entryFields.slice(0, 6).map(f => (
-            <span key={f} className={matchedSet.has(f.toLowerCase()) ? styles.mpillHit : styles.mpillMiss}>
-              {f}{matchedSet.has(f.toLowerCase()) ? ' ✔' : ''}
-            </span>
-          ))}
-          {entryFields.length > 6 && <span className={styles.mpillMiss}>+{entryFields.length - 6}</span>}
-        </div>
-      )}
-      {showFilter && filterFields.length > 0 && (
-        <div className={styles.fieldPillRow}>
-          <Text className={styles.fieldPillLabel}>Filter:</Text>
-          {filterFields.slice(0, 6).map(f => (
-            <span key={f} className={matchedSet.has(f.toLowerCase()) ? styles.mpillHit : styles.mpillMiss}>
-              {f}{matchedSet.has(f.toLowerCase()) ? ' ✔' : ''}
-            </span>
-          ))}
-          {filterFields.length > 6 && <span className={styles.mpillMiss}>+{filterFields.length - 6}</span>}
-        </div>
-      )}
-    </div>
-  );
-}
-
-/* ─────────────────────────────────────────────────────────────────────────
-   Shared small components
-───────────────────────────────────────────────────────────────────────── */
-function DetectionCoverageBanner(): JSX.Element {
-  const styles = useStyles();
-  return (
-    <div className={styles.banner}>
-      <Lightbulb24Regular className={styles.bannerIcon} />
-      <div className={styles.bannerContent}>
-        <div className={styles.bannerTitleRow}>
-          <Text weight="semibold">Detection Coverage Notice</Text>
-          <Badge appearance="filled" shape="rounded" color="warning" size="small">Preview</Badge>
-        </div>
-        <Text as="p" className={styles.bannerListRow}><CloudFlow24Regular className={styles.bannerListIcon} /> <span><strong>Power Automate flows</strong> — cross-entity writes detected from flow JSON definitions.</span></Text>
-        <Text as="p" className={styles.bannerListRow}><ClipboardSettings24Regular className={styles.bannerListIcon} /> <span><strong>Classic Workflows</strong> — cross-entity writes detected from XAML (CreateEntity / UpdateEntity steps).</span></Text>
-        <Text as="p" className={styles.bannerListRow}><ClipboardTaskListLtr24Regular className={styles.bannerListIcon} /> <span><strong>Business Rules (server-scoped)</strong> — server-side rules detected from Dataverse workflow records. Form-scoped (client-only) rules are excluded.</span></Text>
-        <div style={{ marginTop: tokens.spacingVerticalS, paddingTop: tokens.spacingVerticalS, borderTop: `1px solid ${tokens.colorNeutralStroke2}` }}>
-          <div style={{ display: 'flex', alignItems: 'center', gap: tokens.spacingHorizontalXS, marginBottom: tokens.spacingVerticalXXS }}>
-            <Text style={{ fontSize: tokens.fontSizeBase200, color: tokens.colorNeutralForeground3 }}>Coming soon</Text>
-            <Badge appearance="tint" shape="rounded" color="informative" size="small">Planned</Badge>
-          </div>
-          <Text as="p" className={styles.bannerListRow}><BracesVariable24Regular className={styles.bannerListIcon} /> <span><strong>Plugins (deep detection)</strong> — currently shows that a plugin fires (stage, filter attributes, firing status), but cannot identify what the plugin code itself writes to other entities. Plugin assembly decompilation is planned.</span></Text>
-          <Text as="p" className={styles.bannerListRowSpaced}><DocumentGlobe24Regular className={styles.bannerListIcon} /> <span><strong>JavaScript Web Resources (static analysis)</strong> — currently cannot detect cross-entity Dataverse API calls embedded in custom JavaScript. JS static analysis is planned.</span></Text>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function TypeBadge({ type }: { type: AutomationActivation['automationType'] | PipelineStep['automationType'] }): JSX.Element {
-  const color = type === 'Plugin' ? 'important'
-    : type === 'Flow' ? 'success'
-    : type === 'BusinessRule' ? 'brand'
-    : 'warning';
-  const styles = useStyles();
-  return (
-    <Badge appearance="outline" shape="rounded" color={color} className={styles.stepBadge}>
-      {typeIcon(type)} {type}
-    </Badge>
-  );
-}
-
-function OperationBadge({ operation }: { operation: string }): JSX.Element {
-  const styles = useStyles();
-  const label = operation === 'Manual' ? 'On-Demand / Manual Trigger'
-    : operation === 'Action' ? 'Custom Action'
-    : operation;
-  const color = operation === 'Create' ? 'success'
-    : operation === 'Delete' ? 'danger'
-    : operation === 'Manual' ? 'informative'
-    : operation === 'Action' ? 'brand'
-    : 'warning';
-  return (
-    <Badge
-      appearance="tint"
-      shape="rounded"
-      color={color}
-      className={styles.stepBadge}
-    >
-      {label}
-    </Badge>
   );
 }
